Get events from SR Research ASCII data files.

Read data from SR Research ASCII files (samples, fixations, saccades, blinks, etc).

readELascii(file, tStartRE = "TRIALID", tEndRE = "TRIAL_RESULT", subjID = NULL)

Arguments

file

A string giving path/fname to input file (ELascii file).

tStartRE

A string containing a regular expression that uniquely identifies beginnings of trials. It will be the first line for each trial that will be passed to link{getEyelinkTrialData} for processing.

The default value, "TRIALID", is a MSG that occurs immediately before an ET recording block. We use this as the default because it is guaranteed to be present. But, it may not capture information recorded during a trial before that point. A case in point is where a DRIFTCORRECT (drift check) event is present right before the recording block. TRIALID will occur after the drift correct event, meaning that the drift correct offset values captured during the event will not be available. We do not use DRIFTCORRECT as the default value to tStartRE, because it is not guaranteed to be present; not every experimentor chooses to include this event in each trial. Other reasonable choices for this argument may target the EB generated "PREPARE_SEQUENCE" MSG, or even a user generated MSG.

tEndRE

A string containing regular expression that uniquely identifies ends of trials. It will be the last line for each trial that will be passed to getEyelinkTrialData.

The default value, "TRIAL_RESULT", is always the last line to occur in a well-formed trial; the block of "TRIAL_VAR" lines appears right before it. But, if an experiment is aborted prematurely, then the the last trial in the *edf file (and so the *asc file) may not have a proper trial end event for the last trial.

TODO: Test for the case where tStarteRE and TEndRE are mismatched and handle it more gracefully, while throwing a warning.

subjID

If NULL (default), use filename as subject ID. Otherwise use specified string.

Value

List with two elements, one for session information, and one containing a list of trials. Each trial element is itself a list of 6 elements: data.frames enumerating fixations, saccades, blinks, samples, TRIAL_VARs and MSGs for the trial.

Details

SR Research provides a utility (EDF2ASC.exe) that dumps ASCII renderings of their proprietary EDF data file format. This function reads those ASCII files and extracts eye-movement events (fixations, saccades, blinks), specified MSG events, and TRIAL_VARs from them.

Author

Dave Braze davebraze@gmail.com

Monica Li monica.yc.li@gmail.com

Examples

library(FDBeye)

infile <- system.file("extdata/1950006-RAN.asc.gz", package="FDBeye")
D <- readELascii(infile)
#> Warning: seek on a gzfile connection returned an internal error
str(D)
#> List of 2
#>  $ session:'data.frame':	1 obs. of  5 variables:
#>   ..$ subj      : chr "1950006-RAN.asc"
#>   ..$ script    : chr "ET_RAN_wMouse"
#>   ..$ sessdate  : chr "Tue Jan 01 01:01:01 2013"
#>   ..$ srcfile   : chr "1950006.edf"
#>   ..$ conversion: chr "edfapi 3.1 Win32 Mar  2 2011"
#>  $ trials :List of 6
#>   ..$ 0:List of 6
#>   .. ..$ fix     :'data.frame':	64 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:64] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:64] 15293005 15293117 15293253 15293369 15293537 ...
#>   .. .. ..$ etime: num [1:64] 15293085 15293213 15293349 15293509 15294025 ...
#>   .. .. ..$ dur  : num [1:64] 84 100 100 144 492 168 204 240 460 536 ...
#>   .. .. ..$ xpos : num [1:64] 612 538 232 216 313 ...
#>   .. .. ..$ ypos : num [1:64] 482 333 326 304 273 ...
#>   .. .. ..$ pupil: num [1:64] 178 177 182 183 179 188 212 223 219 220 ...
#>   .. ..$ sacc    :'data.frame':	63 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:63] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:63] 15293089 15293217 15293353 15293513 15294029 ...
#>   .. .. ..$ etime  : num [1:63] 15293113 15293249 15293365 15293533 15294069 ...
#>   .. .. ..$ dur    : num [1:63] 28 36 16 24 44 48 24 24 16 24 ...
#>   .. .. ..$ xpos1  : num [1:63] 607 530 230 223 334 ...
#>   .. .. ..$ ypos1  : num [1:63] 473 331 325 308 276 ...
#>   .. .. ..$ xpos2  : num [1:63] 540 227 206 301 672 ...
#>   .. .. ..$ ypos2  : num [1:63] 340 336 298 273 269 ...
#>   .. .. ..$ ampl   : num [1:63] 2.95 5.97 0.71 1.68 6.71 8.26 1.71 1.78 1.19 2.14 ...
#>   .. .. ..$ peakvel: num [1:63] 153 239 79 105 217 261 122 111 106 132 ...
#>   .. ..$ blink   :'data.frame':	4 obs. of  5 variables:
#>   .. .. ..$ event: chr [1:4] "BLINK" "BLINK" "BLINK" "BLINK"
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1
#>   .. .. ..$ stime: num [1:4] 15297485 15303645 15304737 15312945
#>   .. .. ..$ etime: num [1:4] 15297749 15304425 15305413 15313293
#>   .. .. ..$ dur  : num [1:4] 268 784 680 352
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15314331.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "168791.608"
#>   .. .. ..$ AudioRecordingStartTime: chr "147492"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "147797.498098"
#>   .. .. ..$ trial                  : chr "letters1"
#>   .. .. ..$ p1                     : chr "s.png"
#>   .. .. ..$ p2                     : chr "a.png"
#>   .. .. ..$ p3                     : chr "k.png"
#>   .. .. ..$ p4                     : chr "c.png"
#>   .. .. ..$ p5                     : chr "n.png"
#>   .. .. ..$ p6                     : chr "t.png"
#>   .. .. ..$ p7                     : chr "a.png"
#>   .. .. ..$ p8                     : chr "t.png"
#>   .. .. ..$ p9                     : chr "n.png"
#>   .. .. ..$ p10                    : chr "k.png"
#>   .. .. ..$ p11                    : chr "c.png"
#>   .. .. ..$ p12                    : chr "s.png"
#>   .. .. ..$ p13                    : chr "n.png"
#>   .. .. ..$ p14                    : chr "t.png"
#>   .. .. ..$ p15                    : chr "s.png"
#>   .. .. ..$ p16                    : chr "a.png"
#>   .. .. ..$ p17                    : chr "c.png"
#>   .. .. ..$ p18                    : chr "k.png"
#>   .. .. ..$ p19                    : chr "n.png"
#>   .. .. ..$ p20                    : chr "a.png"
#>   .. .. ..$ p21                    : chr "k.png"
#>   .. .. ..$ p22                    : chr "c.png"
#>   .. .. ..$ p23                    : chr "t.png"
#>   .. .. ..$ p24                    : chr "s.png"
#>   .. .. ..$ p25                    : chr "t.png"
#>   .. .. ..$ p26                    : chr "a.png"
#>   .. .. ..$ p27                    : chr "k.png"
#>   .. .. ..$ p28                    : chr "n.png"
#>   .. .. ..$ p29                    : chr "c.png"
#>   .. .. ..$ p30                    : chr "s.png"
#>   .. .. ..$ p31                    : chr "a.png"
#>   .. .. ..$ p32                    : chr "n.png"
#>   .. .. ..$ p33                    : chr "c.png"
#>   .. .. ..$ p34                    : chr "k.png"
#>   .. .. ..$ p35                    : chr "s.png"
#>   .. .. ..$ p36                    : chr "t.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	5529 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:5529] "15293001" "15293005" "15293009" "15293013" ...
#>   .. .. ..$ xpl        : chr [1:5529] "607.6" "605.2" "607.5" "610.9" ...
#>   .. .. ..$ ypl        : chr [1:5529] "477.9" "477.9" "479.1" "481.2" ...
#>   .. .. ..$ psl        : chr [1:5529] "178.0" "178.0" "178.0" "178.0" ...
#>   .. .. ..$ xpr        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:5529] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:5529] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:5529] "-5420.0" "-5421.0" "-5421.0" "-5421.0" ...
#>   .. .. ..$ target_y   : chr [1:5529] "3467.0" "3467.0" "3467.0" "3467.0" ...
#>   .. .. ..$ target_dis : chr [1:5529] "765.0" "764.9" "764.6" "764.4" ...
#>   .. .. ..$ remote_warn: chr [1:5529] "....F........" "....F........" "....F........" "....F........" ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15292938" "15292999" "15292999" "15292999" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 0" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>   ..$ 1:List of 6
#>   .. ..$ fix     :'data.frame':	46 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:46] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:46] 15334225 15334329 15334481 15334689 15335017 ...
#>   .. .. ..$ etime: num [1:46] 15334277 15334453 15334669 15334989 15335317 ...
#>   .. .. ..$ dur  : num [1:46] 56 128 192 304 304 444 196 120 352 276 ...
#>   .. .. ..$ xpos : num [1:46] 620 288 209 301 416 ...
#>   .. .. ..$ ypos : num [1:46] 459 271 278 254 242 ...
#>   .. .. ..$ pupil: num [1:46] 158 168 175 173 175 198 204 199 195 185 ...
#>   .. ..$ sacc    :'data.frame':	45 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:45] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:45] 15334281 15334457 15334673 15334993 15335321 ...
#>   .. .. ..$ etime  : num [1:45] 15334325 15334477 15334685 15335013 15335337 ...
#>   .. .. ..$ dur    : num [1:45] 48 24 16 24 20 24 260 24 24 20 ...
#>   .. .. ..$ xpos1  : num [1:45] 617 281 221 312 438 ...
#>   .. .. ..$ ypos1  : num [1:45] 459 272 274 255 242 ...
#>   .. .. ..$ xpos2  : num [1:45] 284 207 262 400 507 ...
#>   .. .. ..$ ypos2  : num [1:45] 279 278 266 249 237 ...
#>   .. .. ..$ ampl   : num [1:45] 8.12 1.56 0.9 1.87 1.5 1.57 2.68 1.88 1.97 1.82 ...
#>   .. .. ..$ peakvel: num [1:45] 245 106 89 124 104 88 328 123 129 140 ...
#>   .. ..$ blink   :'data.frame':	1 obs. of  5 variables:
#>   .. .. ..$ event: chr "BLINK"
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1
#>   .. .. ..$ stime: num 15336065
#>   .. .. ..$ etime: num 15336245
#>   .. .. ..$ dur  : num 184
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15350360.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "204820.267"
#>   .. .. ..$ AudioRecordingStartTime: chr "188692"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "188993.327098"
#>   .. .. ..$ trial                  : chr "letters2"
#>   .. .. ..$ p1                     : chr "a.png"
#>   .. .. ..$ p2                     : chr "c.png"
#>   .. .. ..$ p3                     : chr "n.png"
#>   .. .. ..$ p4                     : chr "s.png"
#>   .. .. ..$ p5                     : chr "t.png"
#>   .. .. ..$ p6                     : chr "k.png"
#>   .. .. ..$ p7                     : chr "s.png"
#>   .. .. ..$ p8                     : chr "a.png"
#>   .. .. ..$ p9                     : chr "c.png"
#>   .. .. ..$ p10                    : chr "n.png"
#>   .. .. ..$ p11                    : chr "t.png"
#>   .. .. ..$ p12                    : chr "k.png"
#>   .. .. ..$ p13                    : chr "a.png"
#>   .. .. ..$ p14                    : chr "t.png"
#>   .. .. ..$ p15                    : chr "s.png"
#>   .. .. ..$ p16                    : chr "n.png"
#>   .. .. ..$ p17                    : chr "c.png"
#>   .. .. ..$ p18                    : chr "k.png"
#>   .. .. ..$ p19                    : chr "t.png"
#>   .. .. ..$ p20                    : chr "s.png"
#>   .. .. ..$ p21                    : chr "k.png"
#>   .. .. ..$ p22                    : chr "n.png"
#>   .. .. ..$ p23                    : chr "c.png"
#>   .. .. ..$ p24                    : chr "a.png"
#>   .. .. ..$ p25                    : chr "c.png"
#>   .. .. ..$ p26                    : chr "s.png"
#>   .. .. ..$ p27                    : chr "k.png"
#>   .. .. ..$ p28                    : chr "t.png"
#>   .. .. ..$ p29                    : chr "a.png"
#>   .. .. ..$ p30                    : chr "n.png"
#>   .. .. ..$ p31                    : chr "s.png"
#>   .. .. ..$ p32                    : chr "k.png"
#>   .. .. ..$ p33                    : chr "n.png"
#>   .. .. ..$ p34                    : chr "t.png"
#>   .. .. ..$ p35                    : chr "c.png"
#>   .. .. ..$ p36                    : chr "a.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	4231 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:4231] "15334221" "15334225" "15334229" "15334233" ...
#>   .. .. ..$ xpl        : chr [1:4231] "620.0" "619.2" "618.3" "617.7" ...
#>   .. .. ..$ ypl        : chr [1:4231] "458.8" "458.3" "457.6" "456.5" ...
#>   .. .. ..$ psl        : chr [1:4231] "158.0" "158.0" "158.0" "158.0" ...
#>   .. .. ..$ xpr        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:4231] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:4231] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:4231] "-5160.0" "-5160.0" "-5160.0" "-5160.0" ...
#>   .. .. ..$ target_y   : chr [1:4231] "2640.0" "2640.0" "2640.0" "2640.0" ...
#>   .. .. ..$ target_dis : chr [1:4231] "698.1" "698.4" "698.6" "698.8" ...
#>   .. .. ..$ remote_warn: chr [1:4231] "............." "............." "............." "............." ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15334163" "15334217" "15334217" "15334217" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 1" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>   ..$ 2:List of 6
#>   .. ..$ fix     :'data.frame':	47 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:47] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:47] 15357789 15357941 15358041 15358377 15358709 ...
#>   .. .. ..$ etime: num [1:47] 15357889 15358013 15358349 15358685 15358989 ...
#>   .. .. ..$ dur  : num [1:47] 104 76 312 312 284 376 260 276 372 244 ...
#>   .. .. ..$ xpos : num [1:47] 619 356 232 322 415 ...
#>   .. .. ..$ ypos : num [1:47] 460 236 278 253 222 ...
#>   .. .. ..$ pupil: num [1:47] 138 144 148 155 167 189 191 198 199 195 ...
#>   .. ..$ sacc    :'data.frame':	46 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:46] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:46] 15357893 15358017 15358353 15358689 15358993 ...
#>   .. .. ..$ etime  : num [1:46] 15357937 15358037 15358373 15358705 15359009 ...
#>   .. .. ..$ dur    : num [1:46] 48 24 24 20 20 24 24 24 20 28 ...
#>   .. .. ..$ xpos1  : num [1:46] 611 343 231 329 427 ...
#>   .. .. ..$ ypos1  : num [1:46] 454 239 277 246 229 ...
#>   .. .. ..$ xpos2  : num [1:46] 368 255 312 401 496 ...
#>   .. .. ..$ ypos2  : num [1:46] 249 266 265 230 216 ...
#>   .. .. ..$ ampl   : num [1:46] 6.92 1.97 1.75 1.59 1.52 1.98 2 2 1.58 2.11 ...
#>   .. .. ..$ peakvel: num [1:46] 222 144 119 116 123 132 132 136 131 116 ...
#>   .. ..$ blink   :'data.frame':	2 obs. of  5 variables:
#>   .. .. ..$ event: chr [1:2] "BLINK" "BLINK"
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1
#>   .. .. ..$ stime: num [1:2] 15368645 15369657
#>   .. .. ..$ etime: num [1:2] 15368709 15371085
#>   .. .. ..$ dur  : num [1:2] 68 1432
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15371688.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "226147.421"
#>   .. .. ..$ AudioRecordingStartTime: chr "212292"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "212595.721098"
#>   .. .. ..$ trial                  : chr "digits1"
#>   .. .. ..$ p1                     : chr "five.png"
#>   .. .. ..$ p2                     : chr "three.png"
#>   .. .. ..$ p3                     : chr "two.png"
#>   .. .. ..$ p4                     : chr "eight.png"
#>   .. .. ..$ p5                     : chr "seven.png"
#>   .. .. ..$ p6                     : chr "four.png"
#>   .. .. ..$ p7                     : chr "two.png"
#>   .. .. ..$ p8                     : chr "five.png"
#>   .. .. ..$ p9                     : chr "three.png"
#>   .. .. ..$ p10                    : chr "four.png"
#>   .. .. ..$ p11                    : chr "eight.png"
#>   .. .. ..$ p12                    : chr "seven.png"
#>   .. .. ..$ p13                    : chr "two.png"
#>   .. .. ..$ p14                    : chr "four.png"
#>   .. .. ..$ p15                    : chr "five.png"
#>   .. .. ..$ p16                    : chr "three.png"
#>   .. .. ..$ p17                    : chr "eight.png"
#>   .. .. ..$ p18                    : chr "seven.png"
#>   .. .. ..$ p19                    : chr "five.png"
#>   .. .. ..$ p20                    : chr "eight.png"
#>   .. .. ..$ p21                    : chr "four.png"
#>   .. .. ..$ p22                    : chr "seven.png"
#>   .. .. ..$ p23                    : chr "three.png"
#>   .. .. ..$ p24                    : chr "two.png"
#>   .. .. ..$ p25                    : chr "four.png"
#>   .. .. ..$ p26                    : chr "three.png"
#>   .. .. ..$ p27                    : chr "eight.png"
#>   .. .. ..$ p28                    : chr "five.png"
#>   .. .. ..$ p29                    : chr "seven.png"
#>   .. .. ..$ p30                    : chr "two.png"
#>   .. .. ..$ p31                    : chr "eight.png"
#>   .. .. ..$ p32                    : chr "three.png"
#>   .. .. ..$ p33                    : chr "four.png"
#>   .. .. ..$ p34                    : chr "five.png"
#>   .. .. ..$ p35                    : chr "seven.png"
#>   .. .. ..$ p36                    : chr "two.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	3672 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:3672] "15357785" "15357789" "15357793" "15357797" ...
#>   .. .. ..$ xpl        : chr [1:3672] "617.9" "618.3" "618.4" "619.4" ...
#>   .. .. ..$ ypl        : chr [1:3672] "456.3" "456.8" "456.8" "457.3" ...
#>   .. .. ..$ psl        : chr [1:3672] "137.0" "138.0" "138.0" "138.0" ...
#>   .. .. ..$ xpr        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:3672] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:3672] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:3672] "-5281.0" "-5281.0" "-5281.0" "-5281.0" ...
#>   .. .. ..$ target_y   : chr [1:3672] "2631.0" "2631.0" "2632.0" "2632.0" ...
#>   .. .. ..$ target_dis : chr [1:3672] "691.8" "691.8" "691.9" "691.9" ...
#>   .. .. ..$ remote_warn: chr [1:3672] "............." "............." "............." "............." ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15357724" "15357781" "15357781" "15357781" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 2" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>   ..$ 3:List of 6
#>   .. ..$ fix     :'data.frame':	49 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:49] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:49] 15379105 15379269 15379377 15379753 15380405 ...
#>   .. .. ..$ etime: num [1:49] 15379217 15379349 15379729 15380377 15380645 ...
#>   .. .. ..$ dur  : num [1:49] 116 84 356 628 244 304 272 236 288 336 ...
#>   .. .. ..$ xpos : num [1:49] 606 291 219 298 423 ...
#>   .. .. ..$ ypos : num [1:49] 445 314 265 227 216 ...
#>   .. .. ..$ pupil: num [1:49] 134 139 144 149 162 178 183 181 176 169 ...
#>   .. ..$ sacc    :'data.frame':	48 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:48] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:48] 15379221 15379353 15379733 15380381 15380649 ...
#>   .. .. ..$ etime  : num [1:48] 15379265 15379373 15379749 15380401 15380673 ...
#>   .. .. ..$ dur    : num [1:48] 48 24 20 24 28 24 24 24 24 24 ...
#>   .. .. ..$ xpos1  : num [1:48] 600 291 234 324 418 ...
#>   .. .. ..$ ypos1  : num [1:48] 443 308 252 229 214 ...
#>   .. .. ..$ xpos2  : num [1:48] 289 228 290 398 315 ...
#>   .. .. ..$ ypos2  : num [1:48] 322 272 235 214 217 ...
#>   .. .. ..$ ampl   : num [1:48] 7.19 1.57 1.24 1.61 2.2 1.87 2.13 1.62 1.92 1.87 ...
#>   .. .. ..$ peakvel: num [1:48] 233 101 97 94 112 112 145 102 118 109 ...
#>   .. ..$ blink   :'data.frame':	1 obs. of  5 variables:
#>   .. .. ..$ event: chr "BLINK"
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1
#>   .. .. ..$ stime: num 15394517
#>   .. .. ..$ etime: num 15395513
#>   .. .. ..$ dur  : num 1000
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15395632.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "250091.506"
#>   .. .. ..$ AudioRecordingStartTime: chr "233593"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "233897.868098"
#>   .. .. ..$ trial                  : chr "digits2"
#>   .. .. ..$ p1                     : chr "three.png"
#>   .. .. ..$ p2                     : chr "five.png"
#>   .. .. ..$ p3                     : chr "seven.png"
#>   .. .. ..$ p4                     : chr "four.png"
#>   .. .. ..$ p5                     : chr "eight.png"
#>   .. .. ..$ p6                     : chr "two.png"
#>   .. .. ..$ p7                     : chr "eight.png"
#>   .. .. ..$ p8                     : chr "four.png"
#>   .. .. ..$ p9                     : chr "two.png"
#>   .. .. ..$ p10                    : chr "seven.png"
#>   .. .. ..$ p11                    : chr "three.png"
#>   .. .. ..$ p12                    : chr "five.png"
#>   .. .. ..$ p13                    : chr "eight.png"
#>   .. .. ..$ p14                    : chr "five.png"
#>   .. .. ..$ p15                    : chr "three.png"
#>   .. .. ..$ p16                    : chr "seven.png"
#>   .. .. ..$ p17                    : chr "two.png"
#>   .. .. ..$ p18                    : chr "four.png"
#>   .. .. ..$ p19                    : chr "three.png"
#>   .. .. ..$ p20                    : chr "four.png"
#>   .. .. ..$ p21                    : chr "eight.png"
#>   .. .. ..$ p22                    : chr "seven.png"
#>   .. .. ..$ p23                    : chr "two.png"
#>   .. .. ..$ p24                    : chr "five.png"
#>   .. .. ..$ p25                    : chr "four.png"
#>   .. .. ..$ p26                    : chr "two.png"
#>   .. .. ..$ p27                    : chr "three.png"
#>   .. .. ..$ p28                    : chr "five.png"
#>   .. .. ..$ p29                    : chr "seven.png"
#>   .. .. ..$ p30                    : chr "eight.png"
#>   .. .. ..$ p31                    : chr "five.png"
#>   .. .. ..$ p32                    : chr "seven.png"
#>   .. .. ..$ p33                    : chr "four.png"
#>   .. .. ..$ p34                    : chr "eight.png"
#>   .. .. ..$ p35                    : chr "three.png"
#>   .. .. ..$ p36                    : chr "two.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	4329 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:4329] "15379101" "15379105" "15379109" "15379113" ...
#>   .. .. ..$ xpl        : chr [1:4329] "610.5" "609.7" "608.5" "608.0" ...
#>   .. .. ..$ ypl        : chr [1:4329] "446.0" "446.9" "446.7" "446.3" ...
#>   .. .. ..$ psl        : chr [1:4329] "134.0" "134.0" "134.0" "133.0" ...
#>   .. .. ..$ xpr        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:4329] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:4329] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:4329] "-5179.0" "-5179.0" "-5179.0" "-5179.0" ...
#>   .. .. ..$ target_y   : chr [1:4329] "2625.0" "2625.0" "2625.0" "2624.0" ...
#>   .. .. ..$ target_dis : chr [1:4329] "696.4" "696.4" "696.5" "696.5" ...
#>   .. .. ..$ remote_warn: chr [1:4329] "............." "............." "............." "............." ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15379034" "15379098" "15379098" "15379098" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 3" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>   ..$ 4:List of 6
#>   .. ..$ fix     :'data.frame':	48 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:48] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:48] 15401889 15402033 15402121 15402413 15402901 ...
#>   .. .. ..$ etime: num [1:48] 15401989 15402097 15402389 15402873 15403089 ...
#>   .. .. ..$ dur  : num [1:48] 104 68 272 464 192 276 556 556 88 448 ...
#>   .. .. ..$ xpos : num [1:48] 606 354 237 305 405 ...
#>   .. .. ..$ ypos : num [1:48] 453 284 266 236 224 ...
#>   .. .. ..$ pupil: num [1:48] 146 150 161 164 168 177 193 194 184 172 ...
#>   .. ..$ sacc    :'data.frame':	47 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:47] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:47] 15401993 15402101 15402393 15402877 15403093 ...
#>   .. .. ..$ etime  : num [1:47] 15402029 15402117 15402409 15402897 15403113 ...
#>   .. .. ..$ dur    : num [1:47] 40 20 20 24 24 ...
#>   .. .. ..$ xpos1  : num [1:47] 598 345 232 305 399 ...
#>   .. .. ..$ ypos1  : num [1:47] 448 283 261 235 229 ...
#>   .. .. ..$ xpos2  : num [1:47] 375 264 310 386 325 ...
#>   .. .. ..$ ypos2  : num [1:47] 296 275 244 227 235 ...
#>   .. .. ..$ ampl   : num [1:47] 5.81 1.72 1.69 1.73 1.58 1.74 1.13 4.35 1.31 1.77 ...
#>   .. .. ..$ peakvel: num [1:47] 233 148 134 112 92 101 91 271 93 81 ...
#>   .. ..$ blink   :'data.frame':	2 obs. of  5 variables:
#>   .. .. ..$ event: chr [1:2] "BLINK" "BLINK"
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1
#>   .. .. ..$ stime: num [1:2] 15404577 15421349
#>   .. .. ..$ etime: num [1:2] 15405593 15422165
#>   .. .. ..$ dur  : num [1:2] 1020 820
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15422768.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "277227.582"
#>   .. .. ..$ AudioRecordingStartTime: chr "256393"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "256691.838098"
#>   .. .. ..$ trial                  : chr "colors1"
#>   .. .. ..$ p1                     : chr "green.png"
#>   .. .. ..$ p2                     : chr "blue.png"
#>   .. .. ..$ p3                     : chr "red.png"
#>   .. .. ..$ p4                     : chr "yellow.png"
#>   .. .. ..$ p5                     : chr "black.png"
#>   .. .. ..$ p6                     : chr "brown.png"
#>   .. .. ..$ p7                     : chr "green.png"
#>   .. .. ..$ p8                     : chr "brown.png"
#>   .. .. ..$ p9                     : chr "red.png"
#>   .. .. ..$ p10                    : chr "black.png"
#>   .. .. ..$ p11                    : chr "blue.png"
#>   .. .. ..$ p12                    : chr "yellow.png"
#>   .. .. ..$ p13                    : chr "black.png"
#>   .. .. ..$ p14                    : chr "blue.png"
#>   .. .. ..$ p15                    : chr "red.png"
#>   .. .. ..$ p16                    : chr "brown.png"
#>   .. .. ..$ p17                    : chr "yellow.png"
#>   .. .. ..$ p18                    : chr "green.png"
#>   .. .. ..$ p19                    : chr "brown.png"
#>   .. .. ..$ p20                    : chr "blue.png"
#>   .. .. ..$ p21                    : chr "red.png"
#>   .. .. ..$ p22                    : chr "green.png"
#>   .. .. ..$ p23                    : chr "yellow.png"
#>   .. .. ..$ p24                    : chr "black.png"
#>   .. .. ..$ p25                    : chr "black.png"
#>   .. .. ..$ p26                    : chr "blue.png"
#>   .. .. ..$ p27                    : chr "red.png"
#>   .. .. ..$ p28                    : chr "yellow.png"
#>   .. .. ..$ p29                    : chr "green.png"
#>   .. .. ..$ p30                    : chr "brown.png"
#>   .. .. ..$ p31                    : chr "yellow.png"
#>   .. .. ..$ p32                    : chr "red.png"
#>   .. .. ..$ p33                    : chr "blue.png"
#>   .. .. ..$ p34                    : chr "brown.png"
#>   .. .. ..$ p35                    : chr "green.png"
#>   .. .. ..$ p36                    : chr "black.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	5417 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:5417] "15401885" "15401889" "15401893" "15401897" ...
#>   .. .. ..$ xpl        : chr [1:5417] "608.6" "608.2" "606.7" "606.7" ...
#>   .. .. ..$ ypl        : chr [1:5417] "452.6" "451.3" "451.1" "452.0" ...
#>   .. .. ..$ psl        : chr [1:5417] "147.0" "146.0" "146.0" "146.0" ...
#>   .. .. ..$ xpr        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:5417] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:5417] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:5417] "-4961.0" "-4961.0" "-4961.0" "-4961.0" ...
#>   .. .. ..$ target_y   : chr [1:5417] "2625.0" "2625.0" "2625.0" "2625.0" ...
#>   .. .. ..$ target_dis : chr [1:5417] "700.6" "700.6" "700.5" "700.5" ...
#>   .. .. ..$ remote_warn: chr [1:5417] "....F........" "....F........" "....F........" "....F........" ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15401817" "15401881" "15401881" "15401881" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 4" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>   ..$ 5:List of 6
#>   .. ..$ fix     :'data.frame':	52 obs. of  8 variables:
#>   .. .. ..$ event: chr [1:52] "FIX" "FIX" "FIX" "FIX" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime: num [1:52] 15426857 15427005 15427113 15427321 15428153 ...
#>   .. .. ..$ etime: num [1:52] 15426953 15427089 15427293 15428125 15428565 ...
#>   .. .. ..$ dur  : num [1:52] 100 88 184 808 416 ...
#>   .. .. ..$ xpos : num [1:52] 615 317 232 308 421 ...
#>   .. .. ..$ ypos : num [1:52] 450 260 273 250 230 ...
#>   .. .. ..$ pupil: num [1:52] 125 132 136 153 188 189 181 172 168 151 ...
#>   .. ..$ sacc    :'data.frame':	51 obs. of  11 variables:
#>   .. .. ..$ event  : chr [1:51] "SACC" "SACC" "SACC" "SACC" ...
#>   .. .. ..$ eye    : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..$ stime  : num [1:51] 15426957 15427093 15427297 15428129 15428569 ...
#>   .. .. ..$ etime  : num [1:51] 15427001 15427109 15427317 15428149 15428589 ...
#>   .. .. ..$ dur    : num [1:51] 48 20 24 24 24 24 24 24 24 64 ...
#>   .. .. ..$ xpos1  : num [1:51] 606 308 236 312 439 ...
#>   .. .. ..$ ypos1  : num [1:51] 448 263 262 254 234 ...
#>   .. .. ..$ xpos2  : num [1:51] 323 239 308 402 503 ...
#>   .. .. ..$ ypos2  : num [1:51] 266 268 244 238 236 ...
#>   .. .. ..$ ampl   : num [1:51] 7.37 1.48 1.57 1.96 1.37 ...
#>   .. .. ..$ peakvel: num [1:51] 240 121 104 119 87 95 115 117 118 373 ...
#>   .. ..$ blink   :'data.frame':	8 obs. of  5 variables:
#>   .. .. ..$ event: chr [1:8] "BLINK" "BLINK" "BLINK" "BLINK" ...
#>   .. .. ..$ eye  : Factor w/ 1 level "L": 1 1 1 1 1 1 1 1
#>   .. .. ..$ stime: num [1:8] 15434549 15440373 15445469 15446501 15448321 ...
#>   .. .. ..$ etime: num [1:8] 15434609 15440389 15445541 15446673 15448353 ...
#>   .. .. ..$ dur  : num [1:8] 64 20 76 176 36 304 240 88
#>   .. ..$ trialvar:'data.frame':	1 obs. of  43 variables:
#>   .. .. ..$ currentEyeTrackerTime  : chr "15451172.0"
#>   .. .. ..$ currentDisplayPCTime   : chr "305630.46"
#>   .. .. ..$ AudioRecordingStartTime: chr "281343"
#>   .. .. ..$ AlignVoiceTriggerTime  : chr "281644.354098"
#>   .. .. ..$ trial                  : chr "colors2"
#>   .. .. ..$ p1                     : chr "red.png"
#>   .. .. ..$ p2                     : chr "black.png"
#>   .. .. ..$ p3                     : chr "yellow.png"
#>   .. .. ..$ p4                     : chr "green.png"
#>   .. .. ..$ p5                     : chr "blue.png"
#>   .. .. ..$ p6                     : chr "brown.png"
#>   .. .. ..$ p7                     : chr "red.png"
#>   .. .. ..$ p8                     : chr "blue.png"
#>   .. .. ..$ p9                     : chr "yellow.png"
#>   .. .. ..$ p10                    : chr "green.png"
#>   .. .. ..$ p11                    : chr "brown.png"
#>   .. .. ..$ p12                    : chr "black.png"
#>   .. .. ..$ p13                    : chr "blue.png"
#>   .. .. ..$ p14                    : chr "black.png"
#>   .. .. ..$ p15                    : chr "red.png"
#>   .. .. ..$ p16                    : chr "green.png"
#>   .. .. ..$ p17                    : chr "yellow.png"
#>   .. .. ..$ p18                    : chr "brown.png"
#>   .. .. ..$ p19                    : chr "brown.png"
#>   .. .. ..$ p20                    : chr "blue.png"
#>   .. .. ..$ p21                    : chr "red.png"
#>   .. .. ..$ p22                    : chr "black.png"
#>   .. .. ..$ p23                    : chr "yellow.png"
#>   .. .. ..$ p24                    : chr "brown.png"
#>   .. .. ..$ p25                    : chr "yellow.png"
#>   .. .. ..$ p26                    : chr "red.png"
#>   .. .. ..$ p27                    : chr "green.png"
#>   .. .. ..$ p28                    : chr "blue.png"
#>   .. .. ..$ p29                    : chr "green.png"
#>   .. .. ..$ p30                    : chr "black.png"
#>   .. .. ..$ p31                    : chr "yellow.png"
#>   .. .. ..$ p32                    : chr "black.png"
#>   .. .. ..$ p33                    : chr "blue.png"
#>   .. .. ..$ p34                    : chr "red.png"
#>   .. .. ..$ p35                    : chr "green.png"
#>   .. .. ..$ p36                    : chr "brown.png"
#>   .. .. ..$ cell_width             : chr "64"
#>   .. .. ..$ cell_height            : chr "64"
#>   .. ..$ samp    :'data.frame':	6276 obs. of  18 variables:
#>   .. .. ..$ time       : chr [1:6276] "15426853" "15426857" "15426861" "15426865" ...
#>   .. .. ..$ xpl        : chr [1:6276] "616.6" "617.8" "619.1" "618.6" ...
#>   .. .. ..$ ypl        : chr [1:6276] "454.5" "454.0" "454.1" "454.2" ...
#>   .. .. ..$ psl        : chr [1:6276] "126.0" "125.0" "125.0" "125.0" ...
#>   .. .. ..$ xpr        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ ypr        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ psr        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ xvl        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ yvl        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ xvr        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ yvr        : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ xr         : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ yr         : logi [1:6276] NA NA NA NA NA NA ...
#>   .. .. ..$ CR_warn    : chr [1:6276] "..." "..." "..." "..." ...
#>   .. .. ..$ target_x   : chr [1:6276] "-5040.0" "-5040.0" "-5040.0" "-5040.0" ...
#>   .. .. ..$ target_y   : chr [1:6276] "2598.0" "2598.0" "2598.0" "2598.0" ...
#>   .. .. ..$ target_dis : chr [1:6276] "684.3" "684.3" "684.3" "684.3" ...
#>   .. .. ..$ remote_warn: chr [1:6276] "............." "............." "............." "............." ...
#>   .. ..$ msg     :'data.frame':	22 obs. of  3 variables:
#>   .. .. ..$ time   : chr [1:22] "15426790" "15426850" "15426850" "15426850" ...
#>   .. .. ..$ offset : chr [1:22] NA NA NA NA ...
#>   .. .. ..$ message: chr [1:22] "TRIALID 5" "RECCFG CR 250 2 1 L" "ELCLCFG RTABLER" "GAZE_COORDS 0.00 0.00 1279.00 1023.00" ...
#>  - attr(*, "class")= chr [1:2] "ELascii" "list"