Extract a fixation report from an ELascii object

Extract a fixation report from an ELascii object

reportFixations(dat, trialvars = TRUE)

Arguments

dat

An object of class "ELascii", as created by readELascii().

trialvars

Logical, defaults TRUE to include trial variables in fixation report.

Value

A data.frame containing a fixation report with the following columns:

subject: factor for subject ID, taken from the name of the source EDF file.

trialn: factor for trial ID.

time0: integer indicating timestamp (ms) of first sample in recording block.

event: factor for event type, always "FIX".

eye: factor indicating eye for corresponding fixation, ranges over c("L", "R").

stime: integer indicating start time for fixation (ms).

etime: integer indicating end time for fixation (ms).

dur: integer indicating duration of fixation (ms).

xpos: integer, x position of fixation in screen coordinates (pixels).

ypos: integer, y position of fixation in screen coordinates (pixels). Origin is upper left corner of screen.

pupil: integer, pupil size (pixels), may be either area or radius depending on ET settings

Returned data.frame will also include trial_vars from edf file if requested (TRUE by default).

Details

Fixation reports include data from both eyes, if available, as well as subject ID, trial number, time stamp of first sample in each trial. By default, values for trial variables found in each trial are also include, although these may be suppressed.

Author

Dave Braze davebraze@gmail.com

Examples

fname <- system.file("/extdata/1950006-RAN.asc.gz", package="FDBeye")
e <- readELascii(fname)
#> Warning: seek on a gzfile connection returned an internal error
efix <- reportFixations(e)