Extract a trial report from an ELascii object

Extract a trial report from an ELascii object

reportTrials(dat)

Arguments

dat

An object of class "ELascii", as created by readELascii().

Value

Which specific trial variables are available will vary, depending on the idiosyncratic characteristics of the protocol. Regardless, the returned data.frame will contain a report with the following columns, together with any trial variables found in the dataset:

subject: factor for subject ID, taken from the name of the source EDF file.

trialn: factor for trial ID.

time0: integer indicating timestamp (ms) of first sample in recording block.

Returned data.frame will also include trial_vars from edf file if requested (TRUE by default).

Details

trial reports contain summary data from each trial, as well as subject ID, trial number.

Author

Dave Braze davebraze@gmail.com

Examples

fname <- system.file("/extdata/1950006-RAN.asc.gz", package="FDBeye")
e <- readELascii(fname)
#> Warning: seek on a gzfile connection returned an internal error
efix <- reportTrials(e)